C++ client library now compiles on Windows, as do some utilities. Added libreplace code for getopt.

+++ client_libs/libplayerc++/aioproxy.cc
@@ -87,7 +87,7 @@ void
 AioProxy::SetVoltage(uint32_t aIndex, double aVoltage)
 {
   scoped_lock_t lock(mPc->mMutex);
-  playerc_aio_set_output(mDevice, aIndex, aVoltage);
+  playerc_aio_set_output(mDevice, aIndex, static_cast<float> (aVoltage));
 }
 
 std::ostream&
